Remove Rebreather merit perk and update Waterlogged
